Custom Shortcode editor:
- support searching custom fields using min max on an integer field
- js checking of input value for int fields
- fix to show custom filters in the Listings page

+++ lib/shortcode-cpt.php
@@ -356,7 +356,7 @@ class PL_Shortcode_CPT {
 								update_post_meta($id, $key, 'true');
 							}
 							break;
-						case 'numeric':
+						case 'int':
 							if( !empty($args) && !empty($args[$option])) {
 								$args[$option] = (int)$args[$option];
 							}
