+++ lib/shortcodes/search_form.php
@@ -21,8 +21,8 @@ class PL_Form_CPT extends PL_SC_Base {
 
 	protected $options = array(
 		'context'			=> array( 'type' => 'select', 'label' => 'Template', 'default' => ''),
-		'width'				=> array( 'type' => 'numeric', 'label' => 'Width(px)', 'default' => 250 ),
-		'height'			=> array( 'type' => 'numeric', 'label' => 'Height(px)', 'default' => 250 ),
+		'width'				=> array( 'type' => 'int', 'label' => 'Width(px)', 'default' => 250 ),
+		'height'			=> array( 'type' => 'int', 'label' => 'Height(px)', 'default' => 250 ),
 		'widget_class'		=> array( 'type' => 'text', 'label' => 'Widget Class', 'default' => '' ),
 		'ajax'				=> array( 'type' => 'checkbox', 'label' => 'Disable AJAX', 'default' => false ),
 		'formaction'		=> array( 'type' => 'text', 'label' => 'Form URL when AJAX is disabled', 'default' => '' ),
